requirements.txt
langchain
langchain-openai
langchain-core
python-dotenv
langchain-ollama
langchain-chroma
pandas
streamlit